Move Layout component to routes (#959)
Unfortunately, in the current version of React, changing the top-level route to another type (even if the child UI tree is almost the same) will result in React tearing down that tree and building a new one. With this fix, the type of the parent component for each page will stay the same as a user transitions from one page/screen to another. Fixes #960.